Extract swell detection logic to a module
